@@ -1,44 +1,36 @@
 [build-system]
-requires = ["poetry-core"]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
-name = "invoicer"
-version = "0.1.0"
-description = ""
-authors = ["Lukáš Klíma <info@lukasklima.cz>"]
-license = "MIT"
-readme = "README.md"
+package-mode = false
 
 [tool.poetry.dependencies]
-python = "3.11.9"
-fastapi = "^0.111.0"
-pydantic = "^2.7.1"
-sqlalchemy = "^2.0.30"
-alembic = "^1.13.1"
-psycopg2-binary = "^2.9.9"
-psycopg = "^3.1.19"
-psycopg-binary = "^3.1.19"
-python-dotenv = "^1.0.1"
-asyncpg = "^0.29.0"
-pydantic-settings = "^2.2.1"
-weasyprint = "^62.3"
-jinja2 = "^3.1.4"
-qrcode = "^8.0"
-pyjwt = "^2.10.0"
-passlib = {extras = ["bcrypt"], version = "^1.7.4"}
-bcrypt = "4.0.1"
+python = "^3.12"
+fastapi = "^0.116.1"
+pydantic = "^2.11.7"
+sqlalchemy = "^2.0.42"
+alembic = "^1.16.4"
+psycopg2-binary = "^2.9.10"
+python-dotenv = "^1.1.1"
+asyncpg = "^0.30.0"
+pyjwt = "^2.10.1"
+passlib = "^1.7.4"
+bcrypt = "^4.3.0"
+greenlet = "^3.2.3"
+google-auth = "^2.40.3"
+pydantic-settings = "^2.10.1"
+psycopg = "^3.2.10"
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^8.2.1"
-ruff = "^0.4.5"
-mypy = "^1.10.0"
-black = "^24.4.2"
-pre-commit = "^3.7.1"
+pytest = "^8.4.1"
+ruff = "^0.12.7"
+mypy = "^1.17.1"
+pre-commit = "^4.2.0"
 
 
 [tool.poetry.group.production.dependencies]
-uvicorn = "^0.32.1"
+uvicorn = "^0.35.0"
 
 [tool.ruff] # https://beta.ruff.rs/docs/settings/
 fix = true
@@ -84,7 +76,9 @@ lint.select = [
 "migrations/*" = [
   "INP001", # flake8-no-pep420 - checks for PEP 420 violations
   "W291",
+  "S608"
 ]
 
 [tool.ruff.isort]
 force-single-line = true
+
