@@ -19,10 +19,10 @@ if config.config_file_name is not None:
 # for 'autogenerate' support
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
-from core.db import Base  # noqa INP001
-from app.models import * # noqa INP001
+from core.db import BaseModel  # noqa INP001
+from app.users.models import *  # noqa INP001
 
-target_metadata = Base.metadata
+target_metadata = BaseModel.metadata
 load_dotenv()
 
 # other values from the config, defined by the needs of env.py,
